Move spring wrappers into one directory
